+++ sample/index.php
@@ -674,6 +674,25 @@ if (PHP_SAPI == 'cli') {
                     <li class="list-group-item">
                         <div class="row">
                             <div class="col-md-8">
+                                <h5>Get full wallet*
+                                    <small>(Depends on "Create wallet")</small>
+                                </h5>
+                            </div>
+                            <div class="col-md-4">
+                                <a href="wallets/GetFullWalletAsFullAddress.php"
+                                   class="btn btn-primary pull-left execute"> Try
+                                    It <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/wallets/GetFullWalletAsFullAddress.html"
+                                   class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8">
                                 <h5>Add addresses to a wallet*
                                     <small>(Depends on "Create wallet")</small>
                                 </h5>